requirements.txt
# Director AI Video App - Requirements
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1
replicate==0.25.2
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
